server/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.0.1
